+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v6.2.2
+- fix: kanban cards are automatically active if no context has been defined so far.
+- fix: an implicitly active context renders the kanban card implicitly active as well.
+
 # v6.2.1
 - fix: deactivate context uses correct index